requirements.txt
playwright>=1.40.0
groq>=0.4.0
python-dotenv>=1.0.0
yagmail>=0.15.0

